Fixes Recorder System TODO: needs testing, tomorrow :)
- Add DBCharacterRecorder table (ID, CharacterID, Name, IconID, ActionsJson)
- Add RecorderMgr with record/save/delete/rename/import/list/icon/append/insert/discard logic
- Add RecorderActionSpellHandler to execute recorded macro actions in sequence
- Add /recorder player command with full subcommand set
- Load all account character recorder rows at login via single async IsIn query; zero DB reads at list/use time
- Cache rows in AccountRecordersByCharId dict and RecorderDbEntries list on GamePlayer
- Intercept player commands in ScriptMgr while recording; skip /recorder itself
- Enforce MaxActionsPerRecorder (14) and MaxRecordersPerPlayer (20) hard caps
- Enforce MaxRecorderNameLength (24) on save and rename
- Guard StartRecording against overwriting an active unsaved session
- Fix icon-setting spell incorrectly also being recorded after resolving pending icon
- Remove direct icon_id argument from /recorder icon; use next-cast spell icon instead
- Replace all synchronous DB queries on player-own rows with in-memory cache lookups

+++ GameServer/gameutils/ScriptMgr.cs
@@ -271,7 +271,7 @@ namespace DOL.GS
 
                 // Only record player-level commands — never record GM/admin commands.
                 // Checked here (after GuessCommand) so we know the command's privilege level.
-                if (client?.Player != null && myCommand.m_lvl == ePrivLevel.Player &&
+                if (client?.Player != null && myCommand.m_lvl == (uint)ePrivLevel.Player &&
                     (RecorderMgr.IsPlayerRecording(client.Player) || RecorderMgr.HasPendingInsert(client.Player)))
                 {
                     GameCommand recorderCmd = GetCommand("&recorder");
